fix: isolate background jobs, only run with the server
