+++ server/lib/queries.js
@@ -5,9 +5,6 @@ import config from 'config';
 import { memoize, pick } from 'lodash';
 memoize.Cache = Map;
 
-const useCache = ['production', 'staging'].includes(process.env.NODE_ENV);
-const CACHE_REFRESH_INTERVAL = process.env.CACHE_REFRESH_INTERVAL || 1000 * 60 * 60;
-
 /*
 * Hacky way to do currency conversion
 */
@@ -747,26 +744,6 @@ const getCollectivesWithMinBackersQuery = async ({ backerCount = 10, orderBy = '
   return { total, collectives };
 };
 const getCollectivesWithMinBackers = memoize(getCollectivesWithMinBackersQuery, JSON.stringify);
-
-// warming up the cache with the homepage queries
-const cacheEntries = [
-  { method: "getCollectivesOrderedByMonthlySpending", params: {"type":"COLLECTIVE","orderBy":"monthlySpending","orderDirection":"DESC","limit":4,"offset":0,"where":{"type":"COLLECTIVE"}} },
-  { method: "getCollectivesOrderedByMonthlySpending", params: {"type":"ORGANIZATION","orderBy":"monthlySpending","orderDirection":"DESC","limit":6,"offset":0,"where":{"type":"ORGANIZATION"}} },
-  { method: "getCollectivesOrderedByMonthlySpending", params: {"type":"USER","orderBy":"monthlySpending","orderDirection":"DESC","limit":30,"offset":0,"where":{"type":"USER"}} },
-  { method: "getCollectivesWithMinBackers", params: {"type":"COLLECTIVE","isActive":true,"minBackerCount":10,"orderBy":"createdAt","orderDirection":"DESC","limit":4,"offset":0,"where":{"type":"COLLECTIVE","isActive":true}}}
-];
-
-const refreshCache = async () => {
-  Promise.each(cacheEntries, async (entry) => {
-    const res = await queries[`${entry.method}Query`](entry.params);
-    queries[entry.method].cache.set(JSON.stringify(entry.params), res);
-  });
-};
-
-if (useCache) {
-  setInterval(refreshCache, CACHE_REFRESH_INTERVAL);
-  refreshCache();
-}
 
 const queries = {
   getPublicHostsByTotalCollectives,
